third_party/blink/web_tests/external/wpt/web-animations/interfaces/Document/getAnimations-expected.txt

This is a testharness.js-based test.
PASS Test document.getAnimations for non-animated content
PASS Test document.getAnimations for script-generated animations
PASS Test the order of document.getAnimations with script generated animations
PASS Test document.getAnimations for a disconnected node
PASS Test document.getAnimations with null target
FAIL Test document.getAnimations for elements inside same-origin iframes assert_equals: expected 1 but got 0
PASS Triggers a style change event
Harness: the test ran to completion.
